Rename package from .admin to .dashboard
